requirements.txt
requests
python-dotenv
google-generativeai
boto3
urlextract
yt-dlp
qreader
qrcode
pillow